update firmware over uart blank project and script
working firmware updating over uart from a new project UARTFirmware. python script, update_firmware.py, that automatically run commands through the bootloader to update firmware. just need to change the file path of the bin file for the firmware near the bottom of the script. I could change this so that the command for updating and the file path for the firmware can be prompted when the script is run.